Authentication and Authorization

+++ backend/routes/posts.js
@@ -2,6 +2,7 @@ const express = require('express');
 const multer = require('multer');
 
 const Post = require('../models/post');
+const checkAuth = require('../middleware/check-auth');
 
 const router = express.Router();
 
@@ -26,12 +27,13 @@ const storage = multer.diskStorage({
     }
 })
 
-router.post('', multer({storage: storage}).single("image"), (req, res, next) => {
+router.post('', checkAuth,  multer({storage: storage}).single("image"), (req, res, next) => {
     const url = req.protocol + '://' + req.get("host");
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + "/images/" + req.file.filename
+        imagePath: url + "/images/" + req.file.filename,
+        creator: req.userData.userId
     });
     post.save().then(result => {
         res.status(201).json({
@@ -80,7 +82,7 @@ router.get('/:id', (req, res, next) => {
         })
 })
 
-router.put("/:id",  multer({storage: storage}).single("image"), (req, res, next) => {
+router.put("/:id", checkAuth,  multer({storage: storage}).single("image"), (req, res, next) => {
     let imagePath = req.body.imagePath;
     if(req.file) {
         const url = req.protocol + '://' + req.get('host');
@@ -90,22 +92,33 @@ router.put("/:id",  multer({storage: storage}).single("image"), (req, res, next)
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     })
-    Post.updateOne({_id: req.params.id }, post)
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post)
         .then(updatedPost => {
-            res.status(200).json({
-                message: "Update Success"
-            })
+            if(updatedPost.nModified > 0) {
+                res.status(200).json({
+                    message: "Update Success"
+                })
+            } else {
+                res.status(401).json({message: "Not authorized!"})
+            }
+            
         })
 })
 
-router.delete('/:id', (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id })
+router.delete('/:id', checkAuth, (req, res, next) => {
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
         .then(status => {
-            res.status(200).json({
-                message: 'Post Delete Success'
-            })
+            if(status.n > 0) {
+                res.status(200).json({
+                    message: 'Post Delete Success'
+                })
+            } else {
+                res.status(401).json({message: "Not authorized!"}) 
+            }
+           
         })
         .catch(err => console.log(err));
 })
